Fix and improve testing for data module with num_replicas.
