+++ tests/test_locality_Voronoi.py
@@ -2,17 +2,7 @@ import numpy as np
 import numpy.testing as npt
 import freud
 import unittest
-
-
-def sort_rounded_xyz_array(arr, decimals=4):
-    """The order of Voronoi polytopes' vertices is not well-defined. Instead of
-    testing a fixed array, arrays must be sorted by their rounded
-    representations in order to compare their values.
-    """
-    arr = np.asarray(arr)
-    arr = arr.round(decimals)
-    indices = np.lexsort((arr[:, 2], arr[:, 1], arr[:, 0]))
-    return arr[indices]
+from util import sort_rounded_xyz_array
 
 
 class TestVoronoi(unittest.TestCase):
