+++ tests/util.py
@@ -104,6 +104,18 @@ def make_alternating_lattice(lattice_size, angle=0, extra_shell=2):
     return points_1, points_2
 
 
+def sort_rounded_xyz_array(arr, decimals=4):
+    """The order of points returned is not always well-defined, such as in
+    Voronoi or UnitCell creation. Instead of testing a fixed array, arrays must
+    be sorted by their rounded representations in order to compare their
+    values.
+    """
+    arr = np.asarray(arr)
+    arr = arr.round(decimals)
+    indices = np.lexsort((arr[:, 2], arr[:, 1], arr[:, 0]))
+    return arr[indices]
+
+
 def skipIfMissing(library):
     try:
         import importlib
